Redesign Minibrowser
1) Put widgets in window thread
2) webkit events on main thread
